% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asserlist_cleanup.R
\name{asserlist_cleanup}
\alias{asserlist_cleanup}
\title{This program mimics Stata's asserlist_cleanup program. It cleans up the two dataframes assertlist generates and export them to an Excel file.}
\usage{
asserlist_cleanup(
  summary_dataframe,
  output_dataframe,
  filename,
  sheetname,
  fix = FALSE
)
}
\arguments{
\item{summary_dataframe}{the result summary dataframe that assertlist generates}

\item{output_dataframe}{the detailed result dataframe that assertlist generates}

\item{filename}{The name of the Excel file}

\item{sheetname}{The sheetname of the sheet of the Excel file}

\item{fix}{TRUE or FALSE, default to be FALSE. If set to be true, asserlist generates additional columns to help data managers correct (or 'fix') errant data values}
}
\value{
An Excel file that holds the output of assertlist
}
\description{
This program mimics Stata's asserlist_cleanup program. It cleans up the two dataframes assertlist generates and export them to an Excel file.
}
