% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertlist_cleanup.R
\name{assertlist_cleanup}
\alias{assertlist_cleanup}
\title{This program mimics Stata's assertlist_cleanup program. It cleans up the two data frames assertlist generates and exports them to an Excel file.}
\usage{
assertlist_cleanup(
  summary_dataframe,
  output_dataframe,
  filename,
  sheetname,
  fix = FALSE
)
}
\arguments{
\item{summary_dataframe}{the result summary data frame that assertlist generates}

\item{output_dataframe}{the detailed result data frame that assertlist generates}

\item{filename}{The name of the Excel file}

\item{sheetname}{The sheetname of the sheet of the Excel file}

\item{fix}{TRUE or FALSE, default to be FALSE. If set to be true, assertlist generates additional columns to help data managers correct (or 'fix') errant data values}
}
\value{
An Excel file that holds the output of assertlist
}
\description{
This program mimics Stata's assertlist_cleanup program. It cleans up the two data frames assertlist generates and exports them to an Excel file.
}
