% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_covg_category.R
\name{assign_covg_category}
\alias{assign_covg_category}
\title{This program runs coverage analysis to create variables for Proportion of Coverage}
\usage{
assign_covg_category(VCP = "assign_covg_category")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
a dataset
}
\description{
This program runs coverage analysis to create variables for Proportion of Coverage
}
\examples{
assign_covg_category()
}
