% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_MOV_flags.R
\name{calculate_MOV_flags}
\alias{calculate_MOV_flags}
\title{Calculate derived variables for missed opportunity for simultaneous vaccination (MOV) analysis}
\usage{
calculate_MOV_flags(VCP = "calculate_MOV_flags")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
RI_MOV_long_form_data and RI_MOV_flags_to_merge datasets in VCQI_OUTPUT_FOLDER
}
\description{
Calculate derived variables for missed opportunity for simultaneous vaccination (MOV) analysis
}
\examples{
calculate_MOV_flags()
}
