% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_previsit_MOV_flags.R
\name{calculate_previsit_MOV_flags}
\alias{calculate_previsit_MOV_flags}
\title{This program makes one of the datasets needed for later R Shiny visualization}
\usage{
calculate_previsit_MOV_flags(VCP = "calculate_previsit_MOV_flags")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
a dataset
}
\description{
This program makes one of the datasets needed for later R Shiny visualization
}
\examples{
calculate_previsit_MOV_flags()
}
