% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_analysis_metadata_MV.R
\name{check_analysis_metadata_MV}
\alias{check_analysis_metadata_MV}
\title{Check VCQI analysis metadata for missing information and format problems}
\usage{
check_analysis_metadata_MV(VCP = "check_analysis_metadata_MV")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Default run settings, default level4 layout file if needed, errors and/or warnings if conditions not met
}
\description{
Check VCQI analysis metadata for missing information and format problems
}
\examples{
check_analysis_metadata_MV()

}
