% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_HW_analysis_metadata.R
\name{check_HW_analysis_metadata}
\alias{check_HW_analysis_metadata}
\title{Check HW analysis-related globals, datasets and variables}
\usage{
check_HW_analysis_metadata(VCP = "check_HW_analysis_metadata")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Log entries; errors if conditions not met
}
\description{
Check HW analysis-related globals, datasets and variables
}
\examples{
check_HW_analysis_metadata()
}
