% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_miss_VCQI_CM_metadata.R
\name{check_miss_VCQI_CM_metadata}
\alias{check_miss_VCQI_CM_metadata}
\title{Check existence, format, and contents of the cluster metadata (CM) dataset}
\usage{
check_miss_VCQI_CM_metadata(VCP = "check_miss_VCQI_CM_metadata")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Errors and/or warnings if conditions not met
}
\description{
Check existence, format, and contents of the cluster metadata (CM) dataset
}
