% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_RI_COVG_01_03DV.R
\name{check_RI_COVG_01_03DV}
\alias{check_RI_COVG_01_03DV}
\title{Check upstream output from RI_COVG_01}
\usage{
check_RI_COVG_01_03DV(VCP = "check_RI_COVG_01_03DV")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Errors and/or warnings if conditions not met
}
\description{
Check upstream output from RI_COVG_01
}
