% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_RI_schedule_metadata.R
\name{check_RI_schedule_metadata}
\alias{check_RI_schedule_metadata}
\title{Check dose list(s) and vaccination schedule in a VCQI RI run}
\usage{
check_RI_schedule_metadata(VCP = "check_RI_schedule_metadata")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Errors and/or warnings if conditions not met
}
\description{
Check dose list(s) and vaccination schedule in a VCQI RI run
}
\examples{
check_RI_schedule_metadata()
}
