% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_RI_survey_metadata.R
\name{check_RI_survey_metadata}
\alias{check_RI_survey_metadata}
\title{Check the survey-related global values in a VCQI RI run}
\usage{
check_RI_survey_metadata(VCP = "check_RI_survey_metadata")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Errors and/or warnings if conditions not met
}
\description{
Check the survey-related global values in a VCQI RI run
}
\examples{
check_RI_survey_metadata()
}
