% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_RI_dates_and_ticks.R
\name{cleanup_RI_dates_and_ticks}
\alias{cleanup_RI_dates_and_ticks}
\title{Clean up RI analysis related dates and set ticks}
\usage{
cleanup_RI_dates_and_ticks(VCP = "cleanup_RI_dates_and_ticks")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function
name}
}
\value{
Three datasets in VCQI_OUTPUT_FOLDER: <VCQI_RI_DATASET>_clean,
<VCQI_RI_DATASET>_dqd, <VCQI_RI_DATASET>_dq_flags
}
\description{
This program accomplishes several things:
}
\details{
\itemize{
\item It merges the RI and RIHC datasets (if RIHC records were sought)
\item It checks to be sure there are date and tick variables for each dose in the
RI_DOSE_LIST (both in the card and register datasets)
\item If RIHC records were not sought, it puts empty register variables in the
dataset.
\item It assigns a dob_for_valid_dose_calculations, if there are m, d, and y data
elements in the dob from card, history, or register.
\item It looks at the vaccination dates on the cards and in the register,
checking to see that they occurred a) after the child's birthdate, if known,
b) after the earliest possible birthdate in the survey if the child's dob is
not known, c) before the survey began, and d) if a dose in a series, it
checks to be sure that sequential doses occur in chronological order.
\item Where it finds a vaccine date on card or register that does not have these
properties, it sets the date to missing and marks a tick mark instead, so the
child will get credit for having the dose, but the nonsensical date will not
be passed into the measures that interpret vaccination dates.
\item Finally it merges the new dataset containing clean dates and ticks with the
full RI dataset.
}

The dataset that comes from this program will need unique IDs and then it
should have everything needed to calculate the RI analyses.
}
\examples{
cleanup_RI_dates_and_ticks()
}
