% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_VCQI_globals.R
\name{cleanup_VCQI_globals}
\alias{cleanup_VCQI_globals}
\title{Clean up VCQI globals}
\usage{
cleanup_VCQI_globals()
}
\value{
Remove VCQI global values from global environment
}
\description{
Clean up VCQI globals
}
