% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_multi_lingual_plot_title.R
\name{create_multi_lingual_plot_title}
\alias{create_multi_lingual_plot_title}
\title{Check the length of title and split into two lines if two long}
\usage{
create_multi_lingual_plot_title(title_string)
}
\arguments{
\item{title_string}{The original title, a string}
}
\value{
A string
}
\description{
Check the length of title and split into two lines if two long
}
