% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESC_02_00GC.R
\name{DESC_02_00GC}
\alias{DESC_02_00GC}
\title{Check global macros for DESC_02}
\usage{
DESC_02_00GC(VCP = "DESC_02_00GC")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Errors if conditions not met
}
\description{
Check global macros for DESC_02
}
