% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESC_02_01PP_MV.R
\name{DESC_02_01PP_MV}
\alias{DESC_02_01PP_MV}
\title{Pre-process dataset for DESC_02_MV}
\usage{
DESC_02_01PP_MV(VCP = "DESC_02_01PP_MV")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
A dataset (DESC_02_<ANALYSIS_COUNTER>_<DESC_02_COUNTER>)
}
\description{
Pre-process dataset for DESC_02_MV
}
