% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESC_02_03DV.R
\name{DESC_02_03DV}
\alias{DESC_02_03DV}
\title{Calculate derived variables for DESC_02}
\usage{
DESC_02_03DV(VCP = "DESC_02_03DV")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
A dataset (DESC_02_<ANALYSIS_COUNTER>_<DESC_02_COUNTER>)
}
\description{
Calculate derived variables for DESC_02
}
