% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESC_02_04GO.R
\name{DESC_02_04GO}
\alias{DESC_02_04GO}
\title{Generate output databases for DESC_02}
\usage{
DESC_02_04GO(VCP = "DESC_02_04GO")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Databases in VCQI_OUTPUT_FOLDER
}
\description{
Generate output databases for DESC_02
}
