% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESC_02_MV.R
\name{DESC_02_MV}
\alias{DESC_02_MV}
\title{Response to multiple-choice question with only one response}
\usage{
DESC_02_MV(VCP = "DESC_02_MV", cleanup = FALSE)
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Derived variables, databases and tables
}
\description{
Response to multiple-choice question with only one response
}
\examples{
DESC_02_MV()
}
