% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESC_03.R
\name{DESC_03}
\alias{DESC_03}
\title{Response to multiple-choice question where the respondent may select more than one response option.}
\usage{
DESC_03(VCP = "DESC_03", cleanup = FALSE)
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Derived variables, databases and tables
}
\description{
Response to multiple-choice question where the respondent may select more than one response option.
}
\examples{
DESC_03()
}
