% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESC_03_01PP.R
\name{DESC_03_01PP}
\alias{DESC_03_01PP}
\title{Pre-process dataset for DESC_03}
\usage{
DESC_03_01PP(VCP = "DESC_03_01PP")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
A dataset (DESC_03_<ANALYSIS_COUNTER>_<DESC_03_COUNTER>)
}
\description{
Pre-process dataset for DESC_03
}
