% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESC_03_03DV.R
\name{DESC_03_03DV}
\alias{DESC_03_03DV}
\title{Calculate derived variables for DESC_03}
\usage{
DESC_03_03DV(VCP = "DESC_03_03DV")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
A dataset (DESC_03_<ANALYSIS_COUNTER>_<DESC_03_COUNTER>)
}
\description{
Calculate derived variables for DESC_03
}
