% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESC_03_05TOST.R
\name{DESC_03_05TOST}
\alias{DESC_03_05TOST}
\title{Export datasets to Excel for DESC_03}
\usage{
DESC_03_05TOST(VCP = "DESC_03_05TOST")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Sheet(s) in tabular output Excel file in VCQI_OUTPUT_FOLDER
}
\description{
Export datasets to Excel for DESC_03
}
