% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dq_exit_survey.R
\name{dq_exit_survey}
\alias{dq_exit_survey}
\title{This program checks the data to make sure everything aligns with the expected value specified in the protocol document}
\usage{
dq_exit_survey(...)
}
\arguments{
\item{...}{Other arguments}
}
\value{
An Excel file that documents the results of assertlist checks
}
\description{
This program checks the data to make sure everything aligns with the expected value specified in the protocol document
}
\examples{
dq_exit_survey()
}
