% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dq_health_worker.R
\name{dq_health_worker}
\alias{dq_health_worker}
\title{This program pulls in the Health Worker (HW) specified data and runs assertlist checks to identify potential errors prior to running the program}
\usage{
dq_health_worker(...)
}
\arguments{
\item{...}{Other arguments}
}
\value{
An Excel file that documents the results of assertlist checks
}
\description{
This program pulls in the Health Worker (HW) specified data and runs assertlist checks to identify potential errors prior to running the program
}
\examples{
dq_health_worker()
}
