% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ES_STUD_01.R
\name{ES_STUD_01}
\alias{ES_STUD_01}
\title{Invalid doses administered on study day}
\usage{
ES_STUD_01(VCP = "ES_STUD_01")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Derived variables, databases and tables
}
\description{
Invalid doses administered on study day
}
\examples{
ES_STUD_01()
}
