% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ES_STUD_01_00GC.R
\name{ES_STUD_01_00GC}
\alias{ES_STUD_01_00GC}
\title{Check global macros for ES_STUD_01}
\usage{
ES_STUD_01_00GC(VCP = "ES_STUD_01_00GC")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Errors if conditions not met
}
\description{
Check global macros for ES_STUD_01
}
