% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ES_STUD_01_01PP.R
\name{ES_STUD_01_01PP}
\alias{ES_STUD_01_01PP}
\title{Pre-process dataset for ES_STUD_01}
\usage{
ES_STUD_01_01PP(VCP = "ES_STUD_01_01PP")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
A dataset (ES_STUD_01_<ANALYSIS_COUNTER>)
}
\description{
Pre-process dataset for ES_STUD_01
}
