% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ES_STUD_01_06PO.R
\name{ES_STUD_01_06PO}
\alias{ES_STUD_01_06PO}
\title{Make plots for ES_STUD_01}
\usage{
ES_STUD_01_06PO(VCP = "ES_STUD_01_06PO")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Plots in VCQI_OUTPUT_FOLDER
}
\description{
Make plots for ES_STUD_01
}
