% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ES_STUD_02.R
\name{ES_STUD_02}
\alias{ES_STUD_02}
\title{Missed opportunities and valid doses administered on study day}
\usage{
ES_STUD_02(VCP = "ES_STUD_02")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Derived variables, databases and tables
}
\description{
Missed opportunities and valid doses administered on study day
}
\examples{
ES_STUD_02()
}
