% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ES_STUD_02_04GO.R
\name{ES_STUD_02_04GO}
\alias{ES_STUD_02_04GO}
\title{Generate output databases for ES_STUD_02}
\usage{
ES_STUD_02_04GO(VCP = "ES_STUD_02_04GO")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Databases in VCQI_OUTPUT_FOLDER
}
\description{
Generate output databases for ES_STUD_02
}
