% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ES_STUD_03.R
\name{ES_STUD_03}
\alias{ES_STUD_03}
\title{Comparison of invalid, valid and MOVs before and on the study day}
\usage{
ES_STUD_03(VCP = "ES_STUD_03")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Derived variables, databases and tables
}
\description{
Comparison of invalid, valid and MOVs before and on the study day
}
\examples{
ES_STUD_03()
}
