% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ES_STUD_03_03DV.R
\name{ES_STUD_03_03DV}
\alias{ES_STUD_03_03DV}
\title{Calculate derived variables for ES_STUD_03}
\usage{
ES_STUD_03_03DV(VCP = "ES_STUD_03_03DV")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
A dataset (ES_STUD_03_<ANALYSIS_COUNTER>)
}
\description{
Calculate derived variables for ES_STUD_03
}
