% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ES_STUD_03_05TOST.R
\name{ES_STUD_03_05TOST}
\alias{ES_STUD_03_05TOST}
\title{Export datasets to Excel for ES_STUD_03 with long table}
\usage{
ES_STUD_03_05TOST(VCP = "ES_STUD_03_05TOST")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Sheet(s) in tabular output Excel file in VCQI_OUTPUT_FOLDER
}
\description{
Export datasets to Excel for ES_STUD_03 with long table
}
