% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish_unique_HW_ids.R
\name{establish_unique_HW_ids}
\alias{establish_unique_HW_ids}
\title{Establish unique IDs for the HW dataset}
\usage{
establish_unique_HW_ids(VCP = "establish_unique_HW_ids")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
a dataset
}
\description{
Establish unique IDs for the HW dataset
}
\examples{
establish_unique_HW_ids()
}
