% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_table_to_excel.R
\name{export_table_to_excel}
\alias{export_table_to_excel}
\title{Export analysis table(s) to Excel with styles}
\usage{
export_table_to_excel(
  indicator,
  tablename = NULL,
  sheet = NULL,
  brief = TRUE,
  VCP = "export_stable_to_excel"
)
}
\arguments{
\item{indicator}{Name of the analysis indicator}

\item{tablename}{Name of the table}

\item{sheet}{Name of the sheet}

\item{brief}{Produce a brief table or not, default to be TRUE}

\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Sheet(s) with formatted output in an Excel file in VCQI_OUTPUT_FOLDER
}
\description{
Export analysis table(s) to Excel with styles
}
