% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_desc0203_database.R
\name{format_desc0203_database}
\alias{format_desc0203_database}
\title{Pre-process dataset to for DESC_02_05TOST/DESC_03_05TOST}
\usage{
format_desc0203_database(
  VCP = "format_desc0203_database",
  i,
  db,
  wtd,
  indicator,
  datname,
  vid
)
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
datasets
}
\description{
Pre-process dataset to for DESC_02_05TOST/DESC_03_05TOST
}
