% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_es_dv.R
\name{gen_es_dv}
\alias{gen_es_dv}
\title{Reads in the Exit survey dataset and creates the derived variables}
\usage{
gen_es_dv(VCP = "gen_es_dv")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
a dataset
}
\description{
Reads in the Exit survey dataset and creates the derived variables
}
\examples{
gen_es_dv()
}
