% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_hw_dv.R
\name{gen_hw_dv}
\alias{gen_hw_dv}
\title{Reads in the Health Worker survey and creates the derived variables}
\usage{
gen_hw_dv(VCP = "gen_hw_dv")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
a dataset
}
\description{
Reads in the Health Worker survey and creates the derived variables
}
\examples{
gen_hw_dv()
}
