% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_MOV_dvs.R
\name{gen_MOV_dvs}
\alias{gen_MOV_dvs}
\title{This program makes derived variables needed to make study date MOV tables}
\usage{
gen_MOV_dvs(VCP = "gen_MOV_dvs")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
a dataset
}
\description{
This program makes derived variables needed to make study date MOV tables
}
\examples{
gen_MOV_dvs()
}
