% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW_BAR_01.R
\name{HW_BAR_01}
\alias{HW_BAR_01}
\title{Knowledge and Attitude Barriers}
\usage{
HW_BAR_01(VCP = "HW_BAR_01")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Derived variables, databases and tables
}
\description{
Knowledge and Attitude Barriers
}
\examples{
HW_BAR_01()
}
