% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW_BAR_01_01PP.R
\name{HW_BAR_01_01PP}
\alias{HW_BAR_01_01PP}
\title{Pre-process dataset for HW_BAR_01}
\usage{
HW_BAR_01_01PP(VCP = "HW_BAR_01_01PP")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
A dataset (HW_BAR_01_<ANALYSIS_COUNTER>)
}
\description{
Pre-process dataset for HW_BAR_01
}
