% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW_BAR_01_06PO.R
\name{HW_BAR_01_06PO}
\alias{HW_BAR_01_06PO}
\title{Make plots for HW_BAR_01}
\usage{
HW_BAR_01_06PO(VCP = "HW_BAR_01_06PO")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Plots in VCQI_OUTPUT_FOLDER
}
\description{
Make plots for HW_BAR_01
}
