% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW_PRAC_01.R
\name{HW_PRAC_01}
\alias{HW_PRAC_01}
\title{Improper Vaccine Practice Barriers}
\usage{
HW_PRAC_01(VCP = "HW_PRAC_01")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Derived variables, databases and tables
}
\description{
Improper Vaccine Practice Barriers
}
\examples{
HW_PRAC_01()
}
