% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW_PRAC_01_03DV.R
\name{HW_PRAC_01_03DV}
\alias{HW_PRAC_01_03DV}
\title{Calculate derived variables for HW_PRAC_01}
\usage{
HW_PRAC_01_03DV(VCP = "HW_PRAC_01_03DV")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
A dataset (HW_PRAC_01_<ANALYSIS_COUNTER>)
}
\description{
Calculate derived variables for HW_PRAC_01
}
