% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW_PRAC_01_04GO.R
\name{HW_PRAC_01_04GO}
\alias{HW_PRAC_01_04GO}
\title{Generate output databases for HW_PRAC_01}
\usage{
HW_PRAC_01_04GO(VCP = "HW_PRAC_01_04GO")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Databases in VCQI_OUTPUT_FOLDER
}
\description{
Generate output databases for HW_PRAC_01
}
