% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW_PRAC_01_05TOST.R
\name{HW_PRAC_01_05TOST}
\alias{HW_PRAC_01_05TOST}
\title{Export datasets to Excel for HW_PRAC_01}
\usage{
HW_PRAC_01_05TOST(VCP = "HW_PRAC_01_05TOST")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Sheet(s) in tabular output Excel file in VCQI_OUTPUT_FOLDER
}
\description{
Export datasets to Excel for HW_PRAC_01
}
