% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_string.R
\name{language_string}
\alias{language_string}
\title{Get the specific string with the user required language}
\usage{
language_string(language_use, str, replaceq = FALSE)
}
\arguments{
\item{language_use}{The global that stored the language the user choose}

\item{str}{The string to be retrived from the language file}
}
\value{
A string
}
\description{
Get the specific string with the user required language
}
