% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_count_output_database.R
\name{make_count_output_database}
\alias{make_count_output_database}
\title{Create database with estimated counts}
\usage{
make_count_output_database(
  numerator,
  denominator,
  estlabel,
  vid,
  measureid,
  VCP = "make_count_output_database",
  keepnumerator = FALSE
)
}
\arguments{
\item{numerator}{The variable to use as the numerator}

\item{denominator}{The variable to use as the denominator}

\item{estlabel}{Label for the estimate variable}

\item{vid}{Variable id}

\item{measureid}{Analysis indicator id}

\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
A database in VCQI_OUTPUT_FOLDER
}
\description{
Create database with estimated counts
}
