% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table_column.R
\name{make_table_column}
\alias{make_table_column}
\title{Extract a column of data from a database and save it to a table to be exported later}
\usage{
make_table_column(
  tablename,
  dbfilename,
  variable,
  varformat = NA,
  replacevar = NA,
  withvar = NA,
  noannotate = FALSE,
  annovar = NA,
  label = NA,
  annoparenval = 50,
  annosuppval = 25,
  noscale = FALSE
)
}
\arguments{
\item{tablename}{The name of the table (usually TO_indicator name)}

\item{dbfilename}{Value of the database file that holds the column (not a VAD)}

\item{variable}{Name of the variable to put in the column}

\item{varformat}{The format of the variable}

\item{replacevar}{Name of the variable(s) you want to remove/replace}

\item{withvar}{Name of the variable(s) you want to replace with}

\item{noannotate}{Annotate the variable with parentheses or * based on annovar or not, default to be FALSE}

\item{annovar}{Name of the variable to use for deciding whether to annotate the output}

\item{label}{Label of the variable}

\item{annoparenval}{Threshold for annotating results: results based on samples with fewer than annoparenval observations will be surrounded by parentheses (default = 50)}

\item{annosuppval}{Threshold for suppressing results: results based on samples with fewer than annosuppval observations will be suppress (default = 25)}

\item{noscale}{Do not scale the variable to \% (the default value of FALSE means results will be put on the percentage scale)}
}
\value{
A dataset augmented with an additional table column
}
\description{
Extract a column of data from a database and save it to a table to be exported later
}
