% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_unwtd_output_database_MV.R
\name{make_unwtd_output_database_MV}
\alias{make_unwtd_output_database_MV}
\title{Create unweighted database results}
\usage{
make_unwtd_output_database_MV(
  variable,
  estlabel,
  vid,
  measureid,
  VCP = "make_unwtd_output_database_MV",
  keepnumerator = TRUE,
  ...
)
}
\arguments{
\item{variable}{Outcome variable}

\item{estlabel}{Label for the estimate variable}

\item{vid}{Outcome variable ID for filenames}

\item{measureid}{Analysis indicator ID}

\item{VCP}{VCQI current program name to be logged, default to be the function name}

\item{...}{Other arguments}
}
\value{
A database in VCQI_OUTPUT_FOLDER
}
\description{
Create unweighted database results
}
