% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data_for_MOV_dvs.R
\name{merge_data_for_MOV_dvs}
\alias{merge_data_for_MOV_dvs}
\title{This program creates a dataset that will be used to create derived indicator variables for study date MOV tables}
\usage{
merge_data_for_MOV_dvs(VCP = "merge_data_for_MOV_dvs")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
a dataset
}
\description{
This program creates a dataset that will be used to create derived indicator variables for study date MOV tables
}
\examples{
merge_data_for_MOV_dvs()
}
