% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss_vcqi_cleanup.R
\name{miss_vcqi_cleanup}
\alias{miss_vcqi_cleanup}
\title{Exit MISS VCQI gracefully}
\usage{
miss_vcqi_cleanup()
}
\value{
Call to miss_vcqi_halt_immediately
}
\description{
Exit MISS VCQI gracefully
}
\examples{
miss_vcqi_cleanup()
}
