% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_exit_survey.R
\name{pp_exit_survey}
\alias{pp_exit_survey}
\title{This program will check to see which variables are needed for cleanup dates and ticks....}
\usage{
pp_exit_survey(VCP = "pp_exit_survey")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
A dataset and log entries; errors if conditions not met
}
\description{
This program will check to see which variables are needed for cleanup dates and ticks....
}
\examples{
pp_exit_survey()
}
