% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_cleanup.R
\name{preprocess_cleanup}
\alias{preprocess_cleanup}
\title{Clean up dates for doses in SHIFTTO list (and SHIFTFROM list) as required}
\usage{
preprocess_cleanup(s, dat)
}
\arguments{
\item{s}{Source of the vaccination dates (string)}

\item{dat}{Dataset with vaccination date variables}
}
\value{
A list of variable names and a dataset
}
\description{
Clean up dates for doses in SHIFTTO list (and SHIFTFROM list) as required
}
