% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_COVG_01_02_03_04.R
\name{RI_COVG_01_02_03_04}
\alias{RI_COVG_01_02_03_04}
\title{This program runs runs RI_COVG_01, 02, 03 and 04 without exporting}
\usage{
RI_COVG_01_02_03_04(VCP = "RI_COVG_01_02_03_04")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
a dataset
}
\description{
This program runs runs RI_COVG_01, 02, 03 and 04 without exporting
}
