% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_COVG_01_02DQ.R
\name{RI_COVG_01_02DQ}
\alias{RI_COVG_01_02DQ}
\title{Check data quality for RI_COVG_01}
\usage{
RI_COVG_01_02DQ(VCP = "RI_COVG_01_02DQ")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Errors if conditions not met
}
\description{
Check data quality for RI_COVG_01
}
