% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_COVG_01_03DV.R
\name{RI_COVG_01_03DV}
\alias{RI_COVG_01_03DV}
\title{Calculate derived variables for RI_COVG_01}
\usage{
RI_COVG_01_03DV(VCP = "RI_COVG_01_03DV")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
A dataset (RI_COVG_01_<ANALYSIS_COUNTER>)
}
\description{
Calculate derived variables for RI_COVG_01
}
