% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_COVG_02_01PP.R
\name{RI_COVG_02_01PP}
\alias{RI_COVG_02_01PP}
\title{Pre-process dataset for RI_COVG_02}
\usage{
RI_COVG_02_01PP(VCP = "RI_COVG_02_01PP")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
A dataset (RI_COVG_02_<ANALYSIS_COUNTER>)
}
\description{
Pre-process dataset for RI_COVG_02
}
