% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_QUAL_08.R
\name{RI_QUAL_08}
\alias{RI_QUAL_08}
\title{Calculate percent of visits with missed opportunities for simultaneous vaccination (MOV)}
\usage{
RI_QUAL_08(VCP = "RI_QUAL_08")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Databases, tables, and plots
}
\description{
Calculate percent of visits with missed opportunities for simultaneous vaccination (MOV)
}
\examples{
RI_QUAL_08()
}
