% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_QUAL_08_00GC.R
\name{RI_QUAL_08_00GC}
\alias{RI_QUAL_08_00GC}
\title{Check global macros for RI_QUAL_08}
\usage{
RI_QUAL_08_00GC(VCP = "RI_QUAL_08_00GC")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Errors if conditions not met
}
\description{
Check global macros for RI_QUAL_08
}
