% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_QUAL_08_04GO_MV.R
\name{RI_QUAL_08_04GO_MV}
\alias{RI_QUAL_08_04GO_MV}
\title{Generate output databases for RI_QUAL_08}
\usage{
RI_QUAL_08_04GO_MV(VCP = "RI_QUAL_08_04GO_MV")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Databases in VCQI_OUTPUT_FOLDER
}
\description{
Generate output databases for RI_QUAL_08
}
