% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_QUAL_09.R
\name{RI_QUAL_09}
\alias{RI_QUAL_09}
\title{Calculate percent of children with missed opportunities for simultaneous vaccination (MOV)}
\usage{
RI_QUAL_09(VCP = "RI_QUAL_09")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Derived variables, databases, tables, and plots
}
\description{
Calculate percent of children with missed opportunities for simultaneous vaccination (MOV)
}
\examples{
RI_QUAL_09()
}
