% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_QUAL_09_01PP.R
\name{RI_QUAL_09_01PP}
\alias{RI_QUAL_09_01PP}
\title{Pre-process dataset for RI_QUAL_09}
\usage{
RI_QUAL_09_01PP(VCP = "RI_QUAL_09_01PP")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
A dataset (RI_QUAL_09_<ANALYSIS_COUNTER>)
}
\description{
Pre-process dataset for RI_QUAL_09
}
