% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_QUAL_09_03DV.R
\name{RI_QUAL_09_03DV}
\alias{RI_QUAL_09_03DV}
\title{Calculate derived variables for RI_QUAL_09}
\usage{
RI_QUAL_09_03DV(VCP = "RI_QUAL_09_03DV")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
A dataset (RI_QUAL_09_<ANALYSIS_COUNTER>)
}
\description{
Calculate derived variables for RI_QUAL_09
}
