% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_QUAL_09_05TOST.R
\name{RI_QUAL_09_05TOST}
\alias{RI_QUAL_09_05TOST}
\title{Export datasets to Excel for RI_QUAL_09}
\usage{
RI_QUAL_09_05TOST(VCP = "RI_QUAL_09_05TOST")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Sheet(s) in tabular output Excel file in VCQI_OUTPUT_FOLDER
}
\description{
Export datasets to Excel for RI_QUAL_09
}
