% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_RI_dates.R
\name{shift_RI_dates}
\alias{shift_RI_dates}
\title{Shift dose dates in specified dose series}
\usage{
shift_RI_dates(NUM_DOSE_SHIFTS, dat = NA, VCP = "shift_RI_dates")
}
\arguments{
\item{NUM_DOSE_SHIFTS}{Number of dose series for which dose shifting will be done}

\item{dat}{Dataset with vaccination date variables}

\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
A dataset returned to cleanup_RI_dates_and_ticks
}
\description{
Shift dose dates in specified dose series
}
