% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftfromscript.R
\name{shiftfromscript}
\alias{shiftfromscript}
\title{Shift dates from SHIFTFROM to SHIFTTO list}
\usage{
shiftfromscript(s, vlist, dat)
}
\arguments{
\item{s}{Source of the vaccination dates (string)}

\item{vlist}{Names of variables to keep (vector)}

\item{dat}{Dataset with vaccination date variables}
}
\value{
A list of variable names and a dataset
}
\description{
Shift dates from SHIFTFROM to SHIFTTO list
}
