% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftwithinscript.R
\name{shiftwithinscript}
\alias{shiftwithinscript}
\title{Shift dates within a dose list}
\usage{
shiftwithinscript(input, shifttype, s, vlist, dat)
}
\arguments{
\item{input}{A list of doses to shift evidence within}

\item{shifttype}{Name of the dose list (string)}

\item{s}{Source of the vaccination dates (string)}

\item{vlist}{Names of variables to keep (vector)}

\item{dat}{Dataset with vaccination date variables}
}
\value{
A list of variable names and a dataset
}
\description{
Shift dates within a dose list
}
