% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_text.R
\name{split_text}
\alias{split_text}
\title{Split the text with default/customized cutoff point}
\usage{
split_text(text_string, text_cutoff = -99)
}
\arguments{
\item{text_string}{The original title, a string}

\item{text_cutoff}{The cutoff point, an integer}
}
\value{
A string
}
\description{
Split the text with default/customized cutoff point
}
