% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svyp_ci_calc.R
\name{svyp_ci_calc}
\alias{svyp_ci_calc}
\title{Calculate survey-adjusted confidence intervals of different types for an estimated proportion}
\usage{
svyp_ci_calc(
  p,
  stderr,
  n,
  dof,
  level = NULL,
  cilevellist = NULL,
  nweighted = NULL,
  nclusters = NULL,
  method,
  adjust = TRUE,
  truncate = TRUE
)
}
\arguments{
\item{p}{Estimated proportion}

\item{stderr}{Standard error}

\item{n}{Sample size}

\item{dof}{Degrees of freedom}

\item{level}{Confidence level}

\item{cilevellist}{A list of confidence levels (CI calculated for each)}

\item{nweighted}{Pass-through}

\item{nclusters}{Pass-through}

\item{method}{Confidence interval method}

\item{adjust}{If adjust effective sample size, default to be TRUE}

\item{truncate}{Truncate DEFF or not, default to be TRUE}
}
\value{
a dataset
}
\description{
Calculate survey-adjusted confidence intervals of different types for an estimated proportion
}
