% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcqi_basic_fmtids.R
\name{vcqi_basic_fmtids}
\alias{vcqi_basic_fmtids}
\title{Assign default formats for Excel}
\usage{
vcqi_basic_fmtids()
}
\value{
Values that stores Excel styles in the global environment
}
\description{
Assign default formats for Excel
}
