% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcqi_global.R
\name{vcqi_global}
\alias{vcqi_global}
\title{Define and log a value in the global environment}
\usage{
vcqi_global(global, value)
}
\arguments{
\item{global}{The name of the global value}

\item{value}{The value to be assigned}
}
\value{
An R object in the global environment
}
\description{
Define and log a value in the global environment
}
\examples{
vcqi_global(x, 100)
vcqi_global(y, c("a", "b", "c"))

}
