% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcqi_log_comment.R
\name{vcqi_log_comment}
\alias{vcqi_log_comment}
\title{Write comment to VCQI program progress log}
\usage{
vcqi_log_comment(program, level, keyword, comment)
}
\arguments{
\item{program}{The program this comment is associated with}

\item{level}{The level of this comment (1 = high level, e.g. errors, 2 = second-order, e.g. warnings, 3 = general comments, e.g. about files being saved; 4 = comments on properties of datasets, 5 = debugging detail, e.g. program flow comments)}

\item{keyword}{The category this comment belongs to, e.g. Error, Warning, Comment, Data, or Flow}

\item{comment}{The comment to be written}
}
\description{
Write comment to VCQI program progress log
}
\examples{
vcqi_log_comment("Program 1", 1, "Error", "Required value X not defined")
}
