% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcqi_log_global.R
\name{vcqi_log_global}
\alias{vcqi_log_global}
\title{Put the current value of a global into the VCQI log}
\usage{
vcqi_log_global(global)
}
\arguments{
\item{global}{A global value to be logged}
}
\value{
Call to vcqi_log_comment
}
\description{
Put the current value of a global into the VCQI log
}
\examples{
test_value <- 1
vcqi_log_global(test_value)
}
