% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcqi_mdy.R
\name{vcqi_mdy}
\alias{vcqi_mdy}
\title{Parse date from individual month, day, and year components like lubridate::mdy, returning NA if any element is missing}
\usage{
vcqi_mdy(m, d, y)
}
\arguments{
\item{m}{A character or numeric vector indicating month, NAs allowed}

\item{d}{A character or numeric vector indicating day, NAs allowed}

\item{y}{A character or numeric vector indicating year, NAs allowed}
}
\value{
A vector of dates
}
\description{
Parse date from individual month, day, and year components like lubridate::mdy, returning NA if any element is missing
}
\examples{
vcqi_mdy(1,2,2022)
vcqi_mdy(10,NA,2022)

month <- c(1,2,4,5)
date <- c(12,14,16,NA)
year <- rep(2022,4)
vcqi_mdy(month,date,year)
}
