% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcqi_object_exists.R
\name{vcqi_object_exists}
\alias{vcqi_object_exists}
\title{Check if an object is defined, not NULL, not NA, and not an empty vector}
\usage{
vcqi_object_exists(object)
}
\arguments{
\item{object}{Name of object to be tested}
}
\value{
TRUE or FALSE
}
\description{
Check if an object is defined, not NULL, not NA, and not an empty vector
}
\examples{
vcqi_object_exists("test_value")
}
