% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcqi_object_value.R
\name{vcqi_object_value}
\alias{vcqi_object_value}
\title{Check if an object takes on a specified value or is in a range of values}
\usage{
vcqi_object_value(object, value)
}
\arguments{
\item{object}{an object to check}

\item{value}{a value or a range of values}
}
\value{
TRUE or FALSE
}
\description{
Check if an object takes on a specified value or is in a range of values
}
\examples{
X <- 1
vcqi_object_value("X", 1)
vcqi_object_value("X", 2)
}
