% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcqi_read.R
\name{vcqi_read}
\alias{vcqi_read}
\title{Detect file type and read dataset}
\usage{
vcqi_read(file)
}
\arguments{
\item{file}{File to read}
}
\value{
A data frame or tibble
}
\description{
Detect file type and read dataset
}
