% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcqi_rename.R
\name{vcqi_rename}
\alias{vcqi_rename}
\title{Rename a variable, allowing for quoted names}
\usage{
vcqi_rename(dat, old_name, new_name)
}
\arguments{
\item{dat}{Name of dataset to modify}

\item{old_name}{Variable name to change}

\item{new_name}{New variable name to assign}
}
\value{
Modified data frame
}
\description{
Rename a variable, allowing for quoted names
}
\examples{
vcqi_rename(dat, "varname", "newvarname")
}
