% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcqi_to_uwplot.R
\name{vcqi_to_uwplot}
\alias{vcqi_to_uwplot}
\title{Create barplot with unweighted data}
\usage{
vcqi_to_uwplot(
  database,
  title = NULL,
  name = NULL,
  savedata = NA,
  savew = 7,
  saveh = 7,
  VCP = "vcqi_to_uwplot"
)
}
\arguments{
\item{database}{Database used to create the plot}

\item{title}{Title of the plot}

\item{name}{Plot identifier in file name}

\item{savedata}{Path to save the mutated data that created the plot}

\item{savew}{Width of the plot file in inches}

\item{saveh}{Height of the plot file in inches}

\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
A plot
}
\description{
Create barplot with unweighted data
}
