% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertlist_replace.R
\name{assertlist_replace}
\alias{assertlist_replace}
\title{This program mimics Stata's assertlist_replace program. Pulls all populated corrected variable values from a worksheet within an assertlist spreadsheet and produce a new corrected dataset}
\usage{
assertlist_replace(originaldata, excel, sheetname = "checks")
}
\arguments{
\item{originaldata}{the path to the original dataset}

\item{excel}{The path to the Excel file that holds the assertlist_clean output with replace values}

\item{sheetname}{The name of the worksheet that holds the assertlist_clean output with replace values}
}
\value{
A dataset with correted value the user specified
}
\description{
This program mimics Stata's assertlist_replace program. Pulls all populated corrected variable values from a worksheet within an assertlist spreadsheet and produce a new corrected dataset
}
